Refactor existing instruments
